fw_att_control split header into separate file
